Move authority parsing constructor
