+++ grpc/src/client/channel.rs
@@ -258,7 +258,7 @@ impl PersistentChannel {
             .unwrap_or_else(|| resolver_builder.default_authority(&target).to_owned());
         let security_opts = SecurityOpts {
             credentials,
-            authority: parse_authority(&authority),
+            authority: Authority::from_host_port_str(&authority),
             handshake_info: ClientHandshakeInfo::default(),
         };
 
@@ -598,185 +598,5 @@ impl<T: Clone> WatcherIter<T> {
     pub(crate) async fn next(&mut self) -> Option<T> {
         self.rx.changed().await.ok()?;
         Some(self.rx.borrow_and_update().clone())
-    }
-}
-
-/// Parses the host and port from a string. When the input can not be parsed
-/// as (host, port) pair, it returns the entire input as the host.
-fn parse_authority(host_and_port: &str) -> Authority {
-    // Handle bracketed IPv6 addresses (e.g., "[::1]:80").
-    if let Some(stripped) = host_and_port.strip_prefix('[')
-        && let Some((host, port_str)) = stripped.split_once("]:")
-        && let Ok(port) = port_str.parse::<u16>()
-    {
-        return Authority::new(host, Some(port));
-    }
-    // Handle unbracketed addresses (IPv4 or hostnames, e.g., "localhost:8080").
-    if let Some((host, port_str)) = host_and_port.rsplit_once(':')
-        && !host.contains(':')
-        && let Ok(port) = port_str.parse::<u16>()
-    {
-        return Authority::new(host, Some(port));
-    }
-    Authority::new(host_and_port.to_string(), None)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_authority() {
-        struct TestCase {
-            input: &'static str,
-            expected: Authority,
-        }
-
-        let cases = [
-            TestCase {
-                input: "localhost:http",
-                expected: Authority::new("localhost:http", None),
-            },
-            TestCase {
-                input: "localhost:80",
-                expected: Authority::new("localhost", Some(80)),
-            },
-            // host name with zone identifier.
-            TestCase {
-                input: "localhost%lo0:80",
-                expected: Authority::new("localhost%lo0", Some(80)),
-            },
-            TestCase {
-                input: "localhost%lo0:http",
-                expected: Authority::new("localhost%lo0:http", None),
-            },
-            TestCase {
-                input: "[localhost%lo0]:http",
-                expected: Authority::new("[localhost%lo0]:http", None),
-            },
-            TestCase {
-                input: "[localhost%lo0]:80",
-                expected: Authority::new("localhost%lo0", Some(80)),
-            },
-            // IP literal
-            TestCase {
-                input: "127.0.0.1:http",
-                expected: Authority::new("127.0.0.1:http", None),
-            },
-            TestCase {
-                input: "127.0.0.1:80",
-                expected: Authority::new("127.0.0.1", Some(80)),
-            },
-            TestCase {
-                input: "[::1]:http",
-                expected: Authority::new("[::1]:http", None),
-            },
-            TestCase {
-                input: "[::1]:80",
-                expected: Authority::new("::1", Some(80)),
-            },
-            // IP literal with zone identifier.
-            TestCase {
-                input: "[::1%lo0]:http",
-                expected: Authority::new("[::1%lo0]:http", None),
-            },
-            TestCase {
-                input: "[::1%lo0]:80",
-                expected: Authority::new("::1%lo0", Some(80)),
-            },
-            TestCase {
-                input: ":http",
-                expected: Authority::new(":http", None),
-            },
-            TestCase {
-                input: ":80",
-                expected: Authority::new("", Some(80)),
-            },
-            TestCase {
-                input: "grpc.io:",
-                expected: Authority::new("grpc.io:", None),
-            },
-            TestCase {
-                input: "127.0.0.1:",
-                expected: Authority::new("127.0.0.1:", None),
-            },
-            TestCase {
-                input: "[::1]:",
-                expected: Authority::new("[::1]:", None),
-            },
-            TestCase {
-                input: "grpc.io:https%foo",
-                expected: Authority::new("grpc.io:https%foo", None),
-            },
-            TestCase {
-                input: "grpc.io",
-                expected: Authority::new("grpc.io", None),
-            },
-            TestCase {
-                input: "127.0.0.1",
-                expected: Authority::new("127.0.0.1", None),
-            },
-            TestCase {
-                input: "[::1]",
-                expected: Authority::new("[::1]", None),
-            },
-            TestCase {
-                input: "[fe80::1%lo0]",
-                expected: Authority::new("[fe80::1%lo0]", None),
-            },
-            TestCase {
-                input: "[localhost%lo0]",
-                expected: Authority::new("[localhost%lo0]", None),
-            },
-            TestCase {
-                input: "localhost%lo0",
-                expected: Authority::new("localhost%lo0", None),
-            },
-            TestCase {
-                input: "::1",
-                expected: Authority::new("::1", None),
-            },
-            TestCase {
-                input: "fe80::1%lo0",
-                expected: Authority::new("fe80::1%lo0", None),
-            },
-            TestCase {
-                input: "fe80::1%lo0:80",
-                expected: Authority::new("fe80::1%lo0:80", None),
-            },
-            TestCase {
-                input: "[foo:bar]",
-                expected: Authority::new("[foo:bar]", None),
-            },
-            TestCase {
-                input: "[foo:bar]baz",
-                expected: Authority::new("[foo:bar]baz", None),
-            },
-            TestCase {
-                input: "[foo]bar:baz",
-                expected: Authority::new("[foo]bar:baz", None),
-            },
-            TestCase {
-                input: "[foo]:[bar]:baz",
-                expected: Authority::new("[foo]:[bar]:baz", None),
-            },
-            TestCase {
-                input: "[foo]:[bar]baz",
-                expected: Authority::new("[foo]:[bar]baz", None),
-            },
-            TestCase {
-                input: "foo[bar]:baz",
-                expected: Authority::new("foo[bar]:baz", None),
-            },
-            TestCase {
-                input: "foo]bar:baz",
-                expected: Authority::new("foo]bar:baz", None),
-            },
-        ];
-
-        for TestCase { input, expected } in cases {
-            let auth = parse_authority(input);
-            assert_eq!(auth, expected, "authority mismatch for {}", input);
-        }
     }
 }
